version two of the script

+++ scripts/delete_vm.py
@@ -2,17 +2,20 @@
 
 import argparse
 import datetime
+import re
+import sys
 
-#RHEVM
+# RHEVM
 from ovirtsdk.api import API
 
-#VSPHERE
+# VSPHERE
 from psphere.client import Client
 from psphere.managedobjects import VirtualMachine
 
-#RHOS
+# RHOS
 from novaclient.v1_1 import client as novaclient
 
+# utils
 from utils.conf import cfme_data
 from utils.conf import credentials
 from utils.wait import wait_for
@@ -83,6 +86,11 @@ def rhevm_vm_status(api, vm_name):
     return vm.get_status().state
 
 
+def vsphere_vm_status(api, vm_name):
+    vm = VirtualMachine.get(api, name=vm_name)
+    return vm.runtime.powerState
+
+
 def rhevm_is_vm_up(api, vm_name):
     """Returns True if VM on rhevm is up.
 
@@ -144,21 +152,16 @@ def rhevm_delete_vm(api, vm_name):
         vm_name: name of the vm
     """
     vm = api.vms.get(vm_name)
-    if not rhevm_is_vm_down(api, vm_name):
-        if not rhevm_is_vm_up(api, vm_name):
-            #something is wrong, locked image etc
-            print "Something went wrong with %s" % vm_name
-            print "Skipping this vm..."
-            return
+    vm_status = rhevm_vm_status(api, vm_name)
+    if vm_status != 'up' and vm_status != 'down':
+        # something is wrong, locked image etc
+        exc_msg = "VM status: %s" % vm_status
+        raise Exception(exc_msg)
+    if vm_status == "up":
         print "Powering down: %s" % vm.get_name()
         vm.stop()
-        try:
-            wait_for(rhevm_is_vm_down, [api, vm_name], fail_condition=False, delay=10, num_sec=120)
-        except Exception, e:
-            print str(e)
-            print "Trying to continue..."
-            return
-    print "Deleting: %s" % vm.get_name()
+        wait_for(rhevm_is_vm_down, [api, vm_name], fail_condition=False, delay=10, num_sec=120)
+    print "Deleting: %s" % vm_name
     vm.delete()
 
 
@@ -171,21 +174,26 @@ def vsphere_delete_vm(api, vm_name):
         vm_name: name of the vm
     """
     vm = VirtualMachine.get(api, name=vm_name)
-    if not rhevm_is_vm_down(api, vm_name):
-        if not rhevm_is_vm_up(api, vm_name):
-            print "Something went wrong with %s" % vm_name
-            print "Skipping this vm..."
-            return
+    vm_status = vsphere_vm_status(api, vm_name)
+    if vm_status != 'poweredOn' and vm_status != 'poweredOff':
+        #something is wrong, locked image etc
+        exc_msg = "VM status: %s" % vm_status
+        raise Exception(exc_msg)
+    if vm_status == "up":
         print "Powering down: %s" % vm.name
         vm.PowerOffVM_Task()
-        try:
-            wait_for(vsphere_is_vm_down, [api, vm_name], fail_condition=False, delay=10,
-                     num_sec=120)
-        except Exception, e:
-            print str(e)
-            print "Trying to continue..."
-    print "Deleting: %s" % vm.name
+        wait_for(vsphere_is_vm_down, [api, vm_name], fail_condition=False, delay=10,
+                 num_sec=120)
+    print "Deleting: %s" % vm_name
     vm.Destroy_Task()
+
+
+def rhos_delete_vm(api, vm_name):
+    vms = api.servers.list()
+    for vm in vms:
+        if vm.name == vm_name:
+            print "Deleting: %s" % vm_name
+            vm.delete()
 
 
 def is_affected(vm_name, text):
@@ -210,6 +218,7 @@ def rhevm_check_with_api(api, rhevm_use_start_time, run_time, text):
         text: when this string is found in the name of VM, it may be deleted
     """
     vms = api.vms.list()
+    vms_to_delete = []
     for vm in vms:
         vm_name = vm.get_name()
         if not rhevm_use_start_time:
@@ -217,16 +226,17 @@ def rhevm_check_with_api(api, rhevm_use_start_time, run_time, text):
             crt = creation_time.replace(tzinfo=None)
             runtime = TIME_NOW - crt
             if runtime.days >= run_time and is_affected(vm_name, text):
-                print "%s runtime: %s" % (vm_name, runtime.days)
-                rhevm_delete_vm(api, vm_name)
+                print "To delete: %s with runtime: %s" % (vm_name, runtime.days)
+                vms_to_delete.append(vm_name)
         else:
             if rhevm_is_vm_up(api, vm_name):
                 start_time = vm.get_start_time()
                 stt = start_time.replace(tzinfo=None)
                 runtime = TIME_NOW - stt
                 if runtime.days >= run_time and is_affected(vm_name, text):
-                    print "%s runtime: %s" % (vm_name, runtime.days)
-                    rhevm_delete_vm(api, vm_name)
+                    print "To delete: %s with runtime: %s" % (vm_name, runtime.days)
+                    vms_to_delete.append(vm_name)
+    return vms_to_delete
 
 
 def vsphere_check_with_api(api, run_time, text):
@@ -238,12 +248,14 @@ def vsphere_check_with_api(api, run_time, text):
         text: when this string is found in the name of VM, it may be deleted
     """
     vms = VirtualMachine.all(api)
+    vms_to_delete = []
     for vm in vms:
         vm_name = vm.name
         running_for = vm.summary.quickStats.uptimeSeconds / SEC_IN_DAY
         if running_for >= run_time and is_affected(vm_name, text):
-            print "%s runtime: %s" % (vm_name, running_for)
-            vsphere_delete_vm(api, vm_name)
+            print "To delete: %s with runtime: %s" % (vm_name, running_for)
+            vms_to_delete.append(vm_name)
+    return vms_to_delete
 
 
 def rhos_check_with_api(api, run_time, text):
@@ -255,27 +267,125 @@ def rhos_check_with_api(api, run_time, text):
         text: when this string is found in the name of VM, it may be deleted
     """
     vms = api.servers.list()
+    vms_to_delete = []
     for vm in vms:
         vm_name = vm.name
         created = vm.created
-        sep_year = created.find('-')
-        sep_month = created.find('-', sep_year+1)
-        sep_day = created.find('T')
-        sep_hour = created.find(':')
-        sep_minute = created.find(':', sep_hour+1)
-        sep_second = created.find('Z')
-        year = int(created[:sep_year])
-        month = int(created[sep_year+1:sep_month])
-        day = int(created[sep_month+1:sep_day])
-        hour = int(created[sep_day+1:sep_hour])
-        minute = int(created[sep_hour+1:sep_minute])
-        second = int(created[sep_minute+1:sep_second])
-        vm_time = datetime.datetime(year, month, day, hour, minute, second)
+        pattern = re.compile(r'(\d+)-(\d+)-(\d+)\w(\d+):(\d+):(\d+)\w')
+        res = pattern.findall(created)
+        vm_time = datetime.datetime(int(res[0][0]), int(res[0][1]), int(res[0][2]), int(res[0][3]),
+                                    int(res[0][4]), int(res[0][5]))
         runtime = TIME_NOW - vm_time
         if runtime.days >= run_time and is_affected(vm_name, text):
-            print "%s runtime: %s" % (vm_name, runtime.days)
-            print "Deleting: %s" % vm.name
-            vm.delete()
+            print "To delete: %s with runtime: %s" % (vm_name, runtime.days)
+            vms_to_delete.append(vm_name)
+    return vms_to_delete
+
+
+def main(args, providers):
+    was_exception = False
+
+    for provider in providers:
+
+        if 'ec2' in providers[provider]['type']:
+            continue
+
+        print "=== Starting provider %s ===" % provider
+
+        creds = providers[provider]['credentials']
+        username = credentials[creds]['username']
+        password = credentials[creds]['password']
+
+        if 'rhevm' in providers[provider]['type']:
+            rhevm_url = providers[provider]['hostname']
+            # establish api
+            try:
+                rhevm_api = rhevm_create_api(rhevm_url, username, password)
+            except Exception, e:
+                print "RHEVM: Failed to establish connection. Skipping this provider..."
+                print str(e)
+                was_exception = True
+                continue
+            # check for old vms
+            try:
+                delete_vm_list = rhevm_check_with_api(rhevm_api, args.rhevm_use_start_time,
+                                                      args.run_time, args.text)
+            except Exception, e:
+                print "RHEVM: Failed to check for vms using api. Skipping this provider..."
+                print str(e)
+                was_exception = True
+                continue
+            # delete old vms
+            for vm_name in delete_vm_list:
+                try:
+                    pass
+                    rhevm_delete_vm(rhevm_api, vm_name)
+                except Exception, e:
+                    print "RHEVM: Failed to delete vm. Skipping '%s'" % vm_name
+                    print str(e)
+                    was_exception = True
+                    continue
+        if 'virtualcenter' in providers[provider]['type']:
+            vsphere_url = providers[provider]['hostname']
+            # establish api
+            try:
+                vsphere_api = vsphere_create_api(vsphere_url, username, password)
+            except Exception, e:
+                "VSPHERE: Failed to establish connection. Skipping this provider..."
+                print str(e)
+                was_exception = True
+                continue
+            # check for old vms
+            try:
+                delete_vm_list = vsphere_check_with_api(vsphere_api, args.run_time, args.text)
+            except Exception, e:
+                print "VSPHERE: Failed to check for vms using api. Skipping this provider..."
+                print str(e)
+                was_exception = True
+                continue
+            # delete old vms
+            for vm_name in delete_vm_list:
+                try:
+                    pass
+                    vsphere_delete_vm(vsphere_api, vm_name)
+                except Exception, e:
+                    print "VSPHERE: Failed to delete vm. Skipping '%s'" % vm_name
+                    print str(e)
+                    was_exception = True
+                    continue
+        if 'openstack' in providers[provider]['type']:
+            rhos_auth_url = providers[provider]['auth_url']
+            rhos_project_id = credentials[creds]['tenant']
+            # establish api
+            try:
+                rhos_api = rhos_create_api(rhos_auth_url, username, password, rhos_project_id)
+            except Exception, e:
+                print "RHOS: Failed to establish connection. Skipping this provider..."
+                print str(e)
+                was_exception = True
+                continue
+            # check for old vms
+            try:
+                delete_vm_list = rhos_check_with_api(rhos_api, args.run_time, args.text)
+            except Exception, e:
+                print "RHOS: Failed to check for vms using api. Skipping this provider..."
+                print str(e)
+                was_exception = True
+                continue
+            # delete old vms
+            for vm_name in delete_vm_list:
+                try:
+                    pass
+                    rhos_delete_vm(rhos_api, vm_name)
+                except Exception, e:
+                    print "RHOS: Failed to delete vm. Skipping '%s'" % vm_name
+                    print str(e)
+                    was_exception = True
+                    continue
+
+        print "=== End of provider %s ===" % provider
+
+    return was_exception
 
 
 if __name__ == "__main__":
@@ -283,41 +393,4 @@ if __name__ == "__main__":
 
     providers = cfme_data['management_systems']
 
-    for provider in providers:
-        print "=== Starting provider %s ===" % provider
-        if 'rhevm' in providers[provider]['type']:
-            rhevm_credentials = providers[provider]['credentials']
-            rhevm_url = providers[provider]['hostname']
-            rhevm_username = credentials[rhevm_credentials]['username']
-            rhevm_password = credentials[rhevm_credentials]['password']
-            try:
-                rhevm_api = rhevm_create_api(rhevm_url, rhevm_username, rhevm_password)
-                rhevm_check_with_api(rhevm_api, args.rhevm_use_start_time, args.run_time, args.text)
-            except Exception, e:
-                print str(e)
-                print "Skipping this provider..."
-        if 'virtualcenter' in providers[provider]['type']:
-            vsphere_credentials = providers[provider]['credentials']
-            vsphere_url = providers[provider]['hostname']
-            vsphere_username = credentials[vsphere_credentials]['username']
-            vsphere_password = credentials[vsphere_credentials]['password']
-            try:
-                vsphere_api = vsphere_create_api(vsphere_url, vsphere_username, vsphere_password)
-                vsphere_check_with_api(vsphere_api, args.run_time, args.text)
-            except Exception, e:
-                print str(e)
-                print "Skipping this provider..."
-        if 'openstack' in providers[provider]['type']:
-            rhos_auth_url = providers[provider]['auth_url']
-            rhos_credentials = providers[provider]['credentials']
-            rhos_username = credentials[rhos_credentials]['username']
-            rhos_password = credentials[rhos_credentials]['password']
-            rhos_project_id = credentials[rhos_credentials]['tenant']
-            try:
-                rhos_api = rhos_create_api(rhos_auth_url, rhos_username, rhos_password,
-                                           rhos_project_id)
-                rhos_check_with_api(rhos_api, args.run_time, args.text)
-            except Exception, e:
-                print str(e)
-                print "Skipping this provider..."
-        print "=== End of provider %s ===" % provider
+    sys.exit(main(args, providers))
